+++ lib/utilcode/src/main/java/com/blankj/utilcode/util/UtilsActivityLifecycleImpl.java
@@ -6,19 +6,17 @@ import android.app.Application;
 import android.arch.lifecycle.Lifecycle;
 import android.os.Build;
 import android.os.Bundle;
+import android.support.annotation.NonNull;
 import android.util.Log;
 import android.view.Window;
 import android.view.WindowManager;
 
-import org.jetbrains.annotations.NotNull;
-
 import java.lang.reflect.Field;
 import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.Iterator;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
+import java.util.concurrent.ConcurrentHashMap;
 
 /**
  * <pre>
@@ -34,9 +32,8 @@ final class UtilsActivityLifecycleImpl implements Application.ActivityLifecycleC
 
     private final LinkedList<Activity> mActivityList = new LinkedList<>();
 
-    private final List<Utils.OnAppStatusChangedListener>                 mStatusListeners               = new ArrayList<>();
-    private final Map<Activity, List<Utils.OnActivityDestroyedListener>> mDestroyedListenerMap          = new HashMap<>();
-    private final Map<Activity, List<Utils.ActivityLifecycleCallbacks>>  mActivityLifecycleCallbacksMap = new HashMap<>();
+    private final List<Utils.OnAppStatusChangedListener>                mStatusListeners               = new ArrayList<>();
+    private final Map<Activity, List<Utils.ActivityLifecycleCallbacks>> mActivityLifecycleCallbacksMap = new ConcurrentHashMap<>();
 
     private int     mForegroundCount = 0;
     private int     mConfigCount     = 0;
@@ -127,11 +124,36 @@ final class UtilsActivityLifecycleImpl implements Application.ActivityLifecycleC
         }
     }
 
+    private void consumeActivityLifecycleCallbacks(Activity activity, Lifecycle.Event event) {
+        List<Utils.ActivityLifecycleCallbacks> listeners = mActivityLifecycleCallbacksMap.get(activity);
+        if (listeners != null) {
+            for (Utils.ActivityLifecycleCallbacks listener : listeners) {
+                listener.onLifecycleChanged(activity, event);
+                if (event.equals(Lifecycle.Event.ON_CREATE)) {
+                    listener.onActivityCreated(activity);
+                } else if (event.equals(Lifecycle.Event.ON_START)) {
+                    listener.onActivityStarted(activity);
+                } else if (event.equals(Lifecycle.Event.ON_RESUME)) {
+                    listener.onActivityResumed(activity);
+                } else if (event.equals(Lifecycle.Event.ON_PAUSE)) {
+                    listener.onActivityPaused(activity);
+                } else if (event.equals(Lifecycle.Event.ON_STOP)) {
+                    listener.onActivityStopped(activity);
+                } else if (event.equals(Lifecycle.Event.ON_DESTROY)) {
+                    listener.onActivityDestroyed(activity);
+                }
+            }
+            if (event.equals(Lifecycle.Event.ON_DESTROY)) {
+                mActivityLifecycleCallbacksMap.remove(activity);
+            }
+        }
+    }
+
     ///////////////////////////////////////////////////////////////////////////
     // lifecycle start
     ///////////////////////////////////////////////////////////////////////////
     @Override
-    public void onActivityCreated(@NotNull Activity activity, Bundle savedInstanceState) {
+    public void onActivityCreated(@NonNull Activity activity, Bundle savedInstanceState) {
         UtilsBridge.applyLanguage(activity);
         setAnimatorsEnabled();
         setTopActivity(activity);
@@ -139,7 +161,7 @@ final class UtilsActivityLifecycleImpl implements Application.ActivityLifecycleC
     }
 
     @Override
-    public void onActivityStarted(@NotNull Activity activity) {
+    public void onActivityStarted(@NonNull Activity activity) {
         if (!mIsBackground) {
             setTopActivity(activity);
         }
@@ -152,7 +174,7 @@ final class UtilsActivityLifecycleImpl implements Application.ActivityLifecycleC
     }
 
     @Override
-    public void onActivityResumed(@NotNull final Activity activity) {
+    public void onActivityResumed(@NonNull final Activity activity) {
         setTopActivity(activity);
         if (mIsBackground) {
             mIsBackground = false;
@@ -163,7 +185,7 @@ final class UtilsActivityLifecycleImpl implements Application.ActivityLifecycleC
     }
 
     @Override
-    public void onActivityPaused(@NotNull Activity activity) {
+    public void onActivityPaused(@NonNull Activity activity) {
         consumeActivityLifecycleCallbacks(activity, Lifecycle.Event.ON_PAUSE);
     }
 
@@ -183,12 +205,11 @@ final class UtilsActivityLifecycleImpl implements Application.ActivityLifecycleC
     }
 
     @Override
-    public void onActivitySaveInstanceState(@NotNull Activity activity, Bundle outState) {/**/}
+    public void onActivitySaveInstanceState(@NonNull Activity activity, Bundle outState) {/**/}
 
     @Override
-    public void onActivityDestroyed(@NotNull Activity activity) {
+    public void onActivityDestroyed(@NonNull Activity activity) {
         mActivityList.remove(activity);
-        consumeOnActivityDestroyedListener(activity);
         UtilsBridge.fixSoftInputLeaks(activity);
         consumeActivityLifecycleCallbacks(activity, Lifecycle.Event.ON_DESTROY);
     }
@@ -241,49 +262,6 @@ final class UtilsActivityLifecycleImpl implements Application.ActivityLifecycleC
             }
         } else {
             mActivityList.addFirst(activity);
-        }
-    }
-
-    private void consumeOnActivityDestroyedListener(Activity activity) {
-        Iterator<Map.Entry<Activity, List<Utils.OnActivityDestroyedListener>>> iterator
-                = mDestroyedListenerMap.entrySet().iterator();
-        while (iterator.hasNext()) {
-            Map.Entry<Activity, List<Utils.OnActivityDestroyedListener>> entry = iterator.next();
-            if (entry.getKey() == activity) {
-                List<Utils.OnActivityDestroyedListener> value = entry.getValue();
-                for (Utils.OnActivityDestroyedListener listener : value) {
-                    listener.onActivityDestroyed(activity);
-                }
-                iterator.remove();
-            }
-        }
-    }
-
-    private void consumeActivityLifecycleCallbacks(Activity activity, Lifecycle.Event event) {
-        Iterator<Map.Entry<Activity, List<Utils.ActivityLifecycleCallbacks>>> iterator
-                = mActivityLifecycleCallbacksMap.entrySet().iterator();
-        while (iterator.hasNext()) {
-            Map.Entry<Activity, List<Utils.ActivityLifecycleCallbacks>> entry = iterator.next();
-            if (entry.getKey() == activity) {
-                List<Utils.ActivityLifecycleCallbacks> value = entry.getValue();
-                for (Utils.ActivityLifecycleCallbacks listener : value) {
-                    listener.onLifecycleChanged(activity, event);
-                    if (event.equals(Lifecycle.Event.ON_CREATE)) {
-                        listener.onActivityCreated(activity);
-                    } else if (event.equals(Lifecycle.Event.ON_START)) {
-                        listener.onActivityStarted(activity);
-                    } else if (event.equals(Lifecycle.Event.ON_RESUME)) {
-                        listener.onActivityResumed(activity);
-                    } else if (event.equals(Lifecycle.Event.ON_PAUSE)) {
-                        listener.onActivityPaused(activity);
-                    } else if (event.equals(Lifecycle.Event.ON_STOP)) {
-                        listener.onActivityStopped(activity);
-                    } else if (event.equals(Lifecycle.Event.ON_DESTROY)) {
-                        listener.onActivityDestroyed(activity);
-                        iterator.remove();
-                    }
-                }
-            }
         }
     }